alteração de produto

+++ produtos/models.py
@@ -11,7 +11,7 @@ class Produtos(models.Model):
     nome = models.CharField(max_length=50)
     descricao = models.TextField()
     foto = models.ImageField()
-    preco = models.FloatField()
+    preco = models.DecimalField(max_digits=6, decimal_places=2)
     categoria = models.ForeignKey(Categoria, null=True, on_delete=models.SET_NULL)
 
     def __str__(self):
